% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_grid.R
\name{interpolate_to_grid}
\alias{interpolate_to_grid}
\title{Estimate the yearly cycle for one weather station}
\usage{
interpolate_to_grid(station_data, grid_points, variable)
}
\arguments{
\item{station}{name of the station; contains the city name, station number, and station direction}

\item{year}{year in which the expected temperature on each day will be estimated}
}
\value{
a data frame with the following components:
\itemize{
\item \code{type} data frame with column for day number (1-365), and a column for the expected average temperature on each day. 
}
}
\description{
Estimate the expected temperature on each day of the year
for one USCRN/USRCRN weather station.
}
\examples{
#get yearly cycle for weather station Asheville 13 S
station_info<- yearly_cycle("Asheville 13 S")
print(station_info)
}
